Support timeouts for request header parsing (#33)
Real world http servers should most likely have some form of timeouts. The user provided handler does not come into play till after the runtime has had a chance to parse the request headers, so the runtime library should implement a form of timeout that ensures that the runtime only piece runs within an upper bound and the user's error handler is informed if a request header wasn't parsed within this time limit. There is no special support for timeouts for parsing request bodies and/or time duration that a handler needs to run within since those can be managed within the handler because the server context is available to the user, and can be closed at any point to shut down the connection and close the socket.

We should also most likely have a default value for this timeout that is not "infinite". I'm picking 60 seconds for now and this is documented in the mli so should be easy to discover for users creating servers.